+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## NEXT
 ### New Features
+
+- The `Map` class instances now have a `.setTerrainAnimationDuration(d: number)` method
+- The `Map` class instances now have events related to terrain animation `"terrainAnimationStart"` and `"terrainAnimationStop"`
 - expose the function `getWebGLSupportError()` to detect WebGL compatibility
+
 
 ## 2.3.0
 ### Bug Fixes