+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # MapTiler SDK Changelog
 
-## NEXT
+## 2.4.0
 ### New Features
 - Shows a warning message in the map container if WebGL context is lost
 - The event `"webglContextLost"` is now exposed
 - The `Map` class instances now have a `.setTerrainAnimationDuration(d: number)` method
 - The `Map` class instances now have events related to terrain animation `"terrainAnimationStart"` and `"terrainAnimationStop"`
 - expose the function `getWebGLSupportError()` to detect WebGL compatibility
+- Updating to Maplibre v4.7.1
 
 ## 2.3.0
 ### Bug Fixes